chore: minor visual changes feat: load code from code example
