+++ frontend/src/components/Visual/FAQModal/GettingStarted.tsx
@@ -1,4 +1,3 @@
-import { useState } from 'react'
 import {
   Accordion,
   AccordionButton,
@@ -10,57 +9,10 @@ import {
   Code,
   Heading,
   Image,
-  Modal,
-  ModalBody,
-  ModalCloseButton,
-  ModalContent,
-  ModalOverlay,
-  Tab,
-  TabList,
-  TabPanel,
-  TabPanels,
-  Tabs,
   Text,
 } from '@chakra-ui/react'
 
-interface FAQModalProps {
-  open: boolean
-  toggle: () => void
-}
-
-export const FAQModal = (props: FAQModalProps) => {
-  const { open, toggle } = props
-  const [index, setIndex] = useState(0)
-  return (
-    <Modal isOpen={open} onClose={toggle}>
-      <ModalOverlay />
-      <ModalContent maxW="1000px" py={8}>
-        <ModalCloseButton />
-        <ModalBody>
-          <Tabs
-            defaultIndex={index}
-            onChange={(newIndex) => setIndex(newIndex)}
-          >
-            <TabList>
-              <Tab>Getting Started</Tab>
-              <Tab>Code Examples</Tab>
-            </TabList>
-            <TabPanels>
-              <TabPanel>
-                <GettingStarted />
-              </TabPanel>
-              <TabPanel>
-                <CodeExamples />
-              </TabPanel>
-            </TabPanels>
-          </Tabs>
-        </ModalBody>
-      </ModalContent>
-    </Modal>
-  )
-}
-
-const GettingStarted = () => {
+export const GettingStarted = () => {
   return (
     <>
       <Heading as="h3" size="lg">
@@ -271,99 +223,6 @@ const GettingStarted = () => {
           paddingLeft={4}
         />
       </Center>
-    </>
-  )
-}
-
-const dijkstraCode = `\
-from heapq import heappush, heappop
-
-# Number of vertices, number of edges
-n, m = map(int, input().split(' '))
-adj = [[] for _ in range(n)]
-
-# Read edges in the format: vertex vertex weight
-for i in range(m):
-  u, v, w = map(int, input().split(' '))
-  adj[u].append((v, w))
-  adj[v].append((u, w))
-
-distances = [-1 for _ in range(n)]
-distances[0] = 0
-
-# Run Dijkstra's algorithm from node 0 to find
-# shortest distance to all nodes
-pq = []
-heappush(pq, (0, 0))
-while len(pq) > 0:
-  distance, u = heappop(pq)
-  if distances[u] != distance:
-    continue
-  for (v, w) in adj[u]:
-    if distances[v] != -1 and distances[v] < distance + w:
-      continue
-    distances[v] = distance + w
-    heappush(pq, (distances[v], v))`
-
-const dijkstraInput = `\
-5 6
-0 1 10
-1 2 20
-2 3 5
-2 4 15
-3 4 5
-0 4 30`
-
-const CodeExamples = () => {
-  return (
-    <>
-      <Heading as="h3" size="lg" mt={4}>
-        Code Examples
-      </Heading>
-      <Text>Here are some codes and inputs for you to try out!</Text>
-
-      <Accordion defaultIndex={[]} allowMultiple mt={4}>
-        <AccordionItem>
-          <Text as="h4" size="md">
-            <AccordionButton>
-              <Box as="b" flex="1" textAlign="left">
-                Dijkstra's Algorithm (Single Source Shortest Path)
-              </Box>
-              <AccordionIcon />
-            </AccordionButton>
-          </Text>
-          <AccordionPanel pb={4}>
-            <Code display="block" whiteSpace="pre">
-              {dijkstraCode}
-            </Code>
-
-            <Text mt={2}>Input:</Text>
-
-            <Code display="block" whiteSpace="pre">
-              {dijkstraInput}
-            </Code>
-            <br />
-            <Center>
-              <Image
-                src="faq/dijkstra_settings.png"
-                alt="Dijkstra settings"
-                width="50%"
-                display="inline"
-              />
-              <Image
-                src="faq/dijkstra_graph.png"
-                alt="Dijkstra graph"
-                width="35%"
-                display="inline"
-                borderLeft="1px"
-                borderLeftColor="blackAlpha.400"
-                paddingLeft={4}
-              />
-            </Center>
-            <br />
-          </AccordionPanel>
-        </AccordionItem>
-      </Accordion>
     </>
   )
 }
