+++ frontend/src/components/Visual/VisualArea.tsx
@@ -3,6 +3,7 @@ import { AddIcon, InfoIcon } from '@chakra-ui/icons'
 import {
   Button,
   Center,
+  Code,
   Flex,
   Tab,
   TabList,
@@ -67,7 +68,8 @@ export const VisualArea = () => {
           </TabPanels>
         ) : (
           <Center h="full" w="full">
-            Click the "+" button above to add a visualization and get started!
+            Click the <Code>+</Code> button above to add a visualization and get
+            started!
           </Center>
         )}
       </Tabs>
